refactor: separate checkbox-group
